+++ frontend/src/response.ts
@@ -66,11 +66,9 @@ export type ResponseRouteLocation = {
 }
 
 export type Ansiblex = {
-  //_id: string,
   id: number,
   name: string,
-  solarSystemId: bigint,
-  //allianceId: bigint,
+  regionName: string,
 }
 
 export type Waypoint = {